pyproject.toml
[project]
name = "fastapi-react-img-video-upload"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.128.0",
    "fastapi-users[sqlalchemy]>=15.0.3",
    "imagekitio>=5.0.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.40.0",
]
